refactors data fetching to its own file
