+++ src/components/FamilyManifestModal/FamilyManifestModal.js
@@ -1,135 +1,19 @@
 import React from 'react';
 import { get, difference } from 'lodash';
+import Spinner from 'react-spinkit';
+
 import { compose, withProps, withState } from 'recompose';
 import { withFormik } from 'formik';
 import { injectState } from 'freactal/lib/inject';
-import Spinner from 'react-spinkit';
 
-import { withQuery } from '@arranger/components';
-
-import { fileManifestParticipantsAndFamily } from '../../services/downloadData';
 import DataTypeOption from './DataTypeOption';
 import DownloadManifestModal, { DownloadManifestModalFooter } from '../DownloadManifestModal';
 import { ModalSubHeader } from '../Modal';
 import Query from '@arranger/components/dist/Query';
-
-const sqonForDownload = ({ values, familyMemberIds, sqon }) => {
-  const selectedDataTypes = Object.entries(values)
-    .filter(([, val]) => val)
-    .map(([key]) => key);
-  return sqon
-    ? {
-        op: 'or',
-        content: [
-          sqon,
-          {
-            op: 'and',
-            content: [
-              {
-                op: 'in',
-                content: { field: 'data_type', value: selectedDataTypes },
-              },
-              {
-                op: 'in',
-                content: { field: 'participants.kf_id', value: familyMemberIds },
-              },
-            ],
-          },
-        ],
-      }
-    : sqon;
-};
+import { sqonForDownload } from './FamilyManifestModalData';
+import { fileManifestParticipantsAndFamily } from '../../services/downloadData';
 
 const enhance = compose(
-  injectState,
-
-  withQuery(({ sqon, projectId, familyMemberIdAggregation }) => ({
-    renderError: true,
-    projectId,
-    key: 'participantAndFamilyMemberIdsAggregation',
-    query: `
-      query dataTypes($sqon: JSON) {
-        file {
-          aggregations(filters: $sqon) {
-            participants__kf_id {
-              buckets {
-                doc_count
-                key
-              }
-            }
-            participants__family__family_members__kf_id {
-              buckets {
-                doc_count
-                key
-              }
-            }
-          }
-        }
-      }
-    `,
-    variables: { sqon },
-  })),
-  withProps(({ participantAndFamilyMemberIdsAggregation: { data } }) => {
-    const familyMemberIds = (
-      get(data, 'file.aggregations.participants__family__family_members__kf_id.buckets') || []
-    ).map(b => b.key);
-    const participantIds = (get(data, 'file.aggregations.participants__kf_id.buckets') || []).map(
-      b => b.key,
-    );
-    return {
-      familyMemberIds,
-      participantIds,
-      familyMembersWithoutParticipantIds: difference(familyMemberIds, participantIds),
-    };
-  }),
-
-  withQuery(
-    ({
-      sqon,
-      projectId,
-      participantAndFamilyMemberIdsAggregation,
-      familyMembersWithoutParticipantIds,
-    }) => {
-      return {
-        shouldFetch: !participantAndFamilyMemberIdsAggregation.loading,
-        renderError: true,
-        projectId,
-        key: 'dataTypesAggregation',
-        query: `
-          query dataTypes($sqon: JSON) {
-            file {
-              aggregations(filters: $sqon) {
-                data_type {
-                  buckets {
-                    doc_count
-                    key
-                  }
-                }
-              }
-            }
-          }
-        `,
-        variables: {
-          sqon: {
-            op: 'and',
-            content: [
-              {
-                op: 'in',
-                content: {
-                  field: 'participants.kf_id',
-                  value: familyMembersWithoutParticipantIds,
-                },
-              },
-            ],
-          },
-        },
-      };
-    },
-  ),
-  withProps(({ dataTypesAggregation: { data } }) => ({
-    dataTypes: get(data, 'file.aggregations.data_type.buckets') || [],
-  })),
-
   withFormik({
     mapPropsToValues: ({ dataTypes }) =>
       dataTypes.reduce((acc, bucket) => ({ ...acc, [bucket.key]: false }), {}),
@@ -150,7 +34,7 @@ const enhance = compose(
   withState('isDisabled', 'setIsDisabled', false),
 );
 
-const FamilyManifestModal = ({
+const component = ({
   //aggregations
   participantAndFamilyMemberIdsAggregation,
   dataTypesAggregation,
@@ -307,4 +191,4 @@ const FamilyManifestModal = ({
   );
 };
 
-export default enhance(FamilyManifestModal);
+export default enhance(component);
