+++ src/components/FamilyManifestModal/index.js
@@ -1,1 +1,2 @@
 export { default } from './FamilyManifestModal';
+export { default as enhanceWithFamilyModalData } from './FamilyManifestModalData';
